Optimiza tiempos de compilación y mejora el rendimiento general del frontend y API

+++ src/lib/schemas.ts
@@ -13,7 +13,6 @@ const emptyStringToNull = z.preprocess((val) => {
   return val;
 }, z.any());
 
-
 export const modelFormSchema = z.object({
   // --- CAMPOS REQUERIDOS ---
   alias: z.string()
@@ -26,31 +25,34 @@ export const modelFormSchema = z.object({
   // --- CAMPOS OPCIONALES ---
   national_id: emptyStringToNull.pipe(z.string().nullable()),
   gender: emptyStringToNull.pipe(z.enum(['Male', 'Female', 'Other']).nullable()),
-  
-  // SOLUCIÓN: Usamos el preprocesador para convertir "" a null ANTES de validar la fecha.
+
+  // Fechas opcionales
   birth_date: emptyStringToNull.pipe(z.string().nullable()),
   date_joined_agency: emptyStringToNull.pipe(z.string().nullable()),
 
   country: emptyStringToNull.pipe(z.string().nullable()),
-  
-  height_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  shoulders_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  chest_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  bust_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  waist_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  hips_cm: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  shoe_size_eu: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  pants_size: emptyStringToNull.pipe(z.coerce.number().int().positive().nullable()),
-  
+
+  // ✅ Permitir decimales en las medidas
+  height_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  shoulders_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  chest_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  bust_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  waist_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  hips_cm: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  shoe_size_eu: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+  pants_size: emptyStringToNull.pipe(z.coerce.number().positive().nullable()),
+
   top_size: emptyStringToNull.pipe(z.string().nullable()),
   eye_color: emptyStringToNull.pipe(z.string().nullable()),
   hair_color: emptyStringToNull.pipe(z.string().nullable()),
-  
+
   instagram: emptyStringToNull.pipe(z.string().nullable()),
   tiktok: emptyStringToNull.pipe(z.string().nullable()),
 
   email: emptyStringToNull.pipe(z.string().email("El formato del email no es válido.").nullable()),
-  phone_number: emptyStringToNull.pipe(z.string().regex(phoneRegex, "El teléfono debe empezar con + y solo contener números.").nullable()),
+  phone_number: emptyStringToNull.pipe(
+    z.string().regex(phoneRegex, "El teléfono debe empezar con + y solo contener números.").nullable()
+  ),
 
   status: z.enum(['active', 'inactive', 'archived']).default('active'),
 });
